2719 locales are not properly importable in typescript (#2737)
* playing with jsdom

* seems to work

* working on tests

* fixes #2719.

* rollback

* add locales/plugin source and types to npm

* put version back

* update cdn references
