+++ src/docs/partials/installing.html
@@ -13,15 +13,12 @@
     <h2 id='cdn'>Via CDN</h2>
     <pre>
 <code class='language-html'>&lt;!-- Popperjs --&gt;
-&lt;script src=&quot;https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js&quot;
-  crossorigin=&quot;anonymous&quot;&gt;&lt;/script&gt;
+&lt;script src=&quot;https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js&quot; crossorigin=&quot;anonymous&quot;&gt;&lt;/script&gt;
 &lt;!-- Tempus Dominus JavaScript --&gt;
-&lt;script src=&quot;https://cdn.jsdelivr.net/gh/Eonasdan/tempus-dominus@master/dist/js/tempus-dominus.js&quot;
-  crossorigin=&quot;anonymous&quot;&gt;&lt;/script&gt;
+&lt;script src=&quot;https://cdn.jsdelivr.net/npm/@eonasdan/tempus-dominus@6.2.7/dist/js/tempus-dominus.min.js&quot; integrity=&quot;sha256-e5Xw3gmawqYxoKnPHKfgACw7qJ0xRCuz1NK3um0cpho=&quot; crossorigin=&quot;anonymous&quot;&gt;&lt;/script&gt;
 
 &lt;!-- Tempus Dominus Styles --&gt;
-&lt;link href=&quot;https://cdn.jsdelivr.net/gh/Eonasdan/tempus-dominus@master/dist/css/tempus-dominus.css&quot;
-  rel=&quot;stylesheet&quot; crossorigin=&quot;anonymous&quot;&gt;</code>
+&lt;link rel=&quot;stylesheet&quot; href=&quot;https://cdn.jsdelivr.net/npm/@eonasdan/tempus-dominus@6.2.7/dist/css/tempus-dominus.min.css&quot; crossorigin=&quot;anonymous&quot;&gt;</code>
 </pre>
   </div>
