+++ test/test-import.ts
@@ -1,5 +1,9 @@
 import { TempusDominus, version } from '../src/js/tempus-dominus';
+import { localization } from '../src/locales/ru';
 
 const dp: TempusDominus = new TempusDominus(
-  document.getElementById('datetimepicker1')
+  document.getElementById('datetimepicker1'),
+  {
+    localization: localization,
+  }
 );
