+++ test/utilities/serviceLocator.fake.ts
@@ -1,0 +1,17 @@
+// import { afterEach, expect, test, vi } from 'vitest';
+//
+// vi.mock('../../src/js/utilities/service-locator', async () => {
+//   const actual: any = await vi.importActual("../../src/js/utilities/service-locator")
+//   return {
+//     ...actual,
+//     serviceLocator: mockSl
+//     // get serviceLocator() {
+//     //   console.log('getting')
+//     //   return mockSl;
+//     // },
+//     // set serviceLocator(value) {
+//     //   console.log(`setting:`, value)
+//     //   mockSl = value;
+//     // }
+//   };
+// });
