+++ types/locales/nl.d.ts
@@ -1,0 +1,46 @@
+declare const name = 'nl';
+declare const localization: {
+  today: string;
+  clear: string;
+  close: string;
+  selectMonth: string;
+  previousMonth: string;
+  nextMonth: string;
+  selectYear: string;
+  previousYear: string;
+  nextYear: string;
+  selectDecade: string;
+  previousDecade: string;
+  nextDecade: string;
+  previousCentury: string;
+  nextCentury: string;
+  pickHour: string;
+  incrementHour: string;
+  decrementHour: string;
+  pickMinute: string;
+  incrementMinute: string;
+  decrementMinute: string;
+  pickSecond: string;
+  incrementSecond: string;
+  decrementSecond: string;
+  toggleMeridiem: string;
+  selectTime: string;
+  selectDate: string;
+  dayViewHeaderFormat: {
+    month: string;
+    year: string;
+  };
+  locale: string;
+  startOfTheWeek: number;
+  dateFormats: {
+    LT: string;
+    LTS: string;
+    L: string;
+    LL: string;
+    LLL: string;
+    LLLL: string;
+  };
+  ordinal: (n: any) => string;
+  format: string;
+};
+export { localization, name };
