+++ types/plugins/customDateFormat.d.ts
@@ -1,0 +1,2 @@
+declare const _default: (_: any, tdClasses: any) => void;
+export { _default as default };
