+++ types/plugins/fa-five.d.ts
@@ -1,0 +1,14 @@
+declare const faFiveIcons: {
+  type: string;
+  time: string;
+  date: string;
+  up: string;
+  down: string;
+  previous: string;
+  next: string;
+  today: string;
+  clear: string;
+  close: string;
+};
+declare const load: (_: any, __: any, tdFactory: any) => void;
+export { faFiveIcons, load };
