+++ types/plugins/moment-parse.d.ts
@@ -1,0 +1,2 @@
+declare const _default: (option: any, tdClasses: any, tdFactory: any) => void;
+export { _default as default };
