--- types/utilities/actionTypes.d.ts
@@ -1,27 +1,0 @@
-export declare enum ActionTypes {
-  next = 'next',
-  previous = 'previous',
-  pickerSwitch = 'pickerSwitch',
-  selectMonth = 'selectMonth',
-  selectYear = 'selectYear',
-  selectDecade = 'selectDecade',
-  selectDay = 'selectDay',
-  selectHour = 'selectHour',
-  selectMinute = 'selectMinute',
-  selectSecond = 'selectSecond',
-  incrementHours = 'incrementHours',
-  incrementMinutes = 'incrementMinutes',
-  incrementSeconds = 'incrementSeconds',
-  decrementHours = 'decrementHours',
-  decrementMinutes = 'decrementMinutes',
-  decrementSeconds = 'decrementSeconds',
-  toggleMeridiem = 'toggleMeridiem',
-  togglePicker = 'togglePicker',
-  showClock = 'showClock',
-  showHours = 'showHours',
-  showMinutes = 'showMinutes',
-  showSeconds = 'showSeconds',
-  clear = 'clear',
-  close = 'close',
-  today = 'today',
-}
