--- types/utilities/calendar-modes.d.ts
@@ -1,9 +1,0 @@
-import { Unit } from '../datetime';
-import ViewMode from './view-mode';
-declare const CalendarModes: {
-  name: keyof ViewMode;
-  className: string;
-  unit: Unit;
-  step: number;
-}[];
-export default CalendarModes;
