--- types/utilities/default-options.d.ts
@@ -1,3 +1,0 @@
-import Options from './options';
-declare const DefaultOptions: Options;
-export default DefaultOptions;
