--- types/utilities/view-mode.d.ts
@@ -1,8 +1,0 @@
-type ViewMode = {
-  clock: any;
-  calendar: any;
-  months: any;
-  years: any;
-  decades: any;
-};
-export default ViewMode;
